% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.r
\name{makebiolik}
\alias{makebiolik}
\title{Biological Two-Separator Likelihood}
\usage{
makebiolik(a = 0.2)
}
\arguments{
\item{a}{Separation parameter controlling the spacing between the two logits.}
}
\value{
A function mapping a latent vector \code{z} to a probability vector.
}
\description{
Same structure as \code{makeneurolik()}, but with a different default separation
parameter. This can be used to model simple biological switching systems
or coarse-grained gene expression states.
}
