% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{or_or}
\alias{or_or}
\alias{\%||\%}
\title{Default Operator \%||\%}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Primary value.}

\item{b}{Default value.}
}
\value{
Either \code{a} or \code{b}.
}
\description{
Returns \code{b} if \code{a} is NULL, otherwise returns \code{a}.
}
\keyword{internal}
