% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rareflow-package}
\alias{rareflow}
\alias{rareflow-package}
\title{rareflow: Rare-event modeling with normalizing flows}
\description{
Tools for variational inference, rare-event estimation,
Girsanov tilting, and Freidlin–Wentzell quasi-potentials.
}
\author{
\strong{Maintainer}: Pietro Piu \email{pietro.piu.si@gmail.com}

}
\keyword{internal}
