% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sanovprob}
\alias{sanovprob}
\title{Sanov Probability Bound}
\usage{
sanovprob(Q, P, n)
}
\arguments{
\item{Q}{Observed empirical distribution.}

\item{P}{True distribution.}

\item{n}{Sample size.}
}
\value{
A numeric upper bound.
}
\description{
Computes the classical Sanov upper bound:
}
\details{
\deqn{P(Q_n \approx Q) \le \exp\{-n \, KL(Q \| P)\}}

where Q is the empirical distribution and P is the true distribution.
}
