% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.r, R/utils.r
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Activation}
\usage{
sigmoid(x)

sigmoid(x)
}
\arguments{
\item{x}{Numeric input.}
}
\value{
Numeric value between 0 and 1.

Numeric output in (0,1).
}
\description{
Standard logistic sigmoid function.

Standard logistic sigmoid function.
}
\keyword{internal}
