% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{softmax}
\alias{softmax}
\title{Softmax Function}
\usage{
softmax(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Probability vector summing to 1.
}
\description{
Computes a numerically stable softmax.
}
\keyword{internal}
